+++ base/sfem_macros.h
@@ -1,0 +1,7 @@
+#ifndef SFEM_MACROS_H
+#define SFEM_MACROS_H
+
+#define MIN(a, b) ((a) < (b) ? (a) : (b))
+#define MAX(a, b) ((a) > (b) ? (a) : (b))
+
+#endif //SFEM_MACROS_H
